+++ src/ops/list_objects.rs
@@ -2,7 +2,7 @@
 
 use crate::error::S3Result;
 use crate::error_code::S3ErrorCode;
-use crate::header::names::X_AMZ_REQUEST_PAYER;
+use crate::headers::names::X_AMZ_REQUEST_PAYER;
 use crate::output::{wrap_output, S3Output, XmlErrorResponse};
 use crate::query::GetQuery;
 use crate::utils::{RequestExt, ResponseExt, XmlWriterExt};
